feat: Split Parser into multiple files with multidispatch
